+++ src/bootstrap/plugins/filter.plugin.ts
@@ -1,8 +1,8 @@
 import Vue from 'vue'
-import { FilterService } from '~/shared/utils/filter.service'
+import filters from '~/shared/filters'
 
 export default {
   install() {
-    Vue.prototype.$filter = FilterService
+    Vue.prototype.$filter = filters
   }
 }