+++ src/pages/demos/calender.vue
@@ -1,7 +1,6 @@
 <template>
   <page-container>
     <a-card>
-      <a-button @click="onClick">123</a-button>
       <FullCalendar
         defaultView="dayGridMonth"
         :plugins="calendarPlugins"
@@ -9,6 +8,7 @@
         :weekends="false"
         :locale="$app.state.locale"
         :buttonText="buttonText"
+        @eventClick="onEventClick"
       />
     </a-card>
   </page-container>
@@ -34,10 +34,10 @@ export default class Calender extends Vue {
   private readonly calendarPlugins = [dayGridPlugin]
 
   private events = [
-    { title: '今天有点事情做', date: '2019-12-30' },
-    { title: '这天好像也有点', date: '2019-12-31' },
-    { title: '中午需要点外卖', date: '2020-01-02' },
-    { title: '下午有空开个会', date: '2020-01-02' }
+    { id: '1', title: '今天有点事情做', date: '2019-12-30' },
+    { id: '2', title: '这天好像也有点', date: '2019-12-31' },
+    { id: '3', title: '中午需要点外卖', date: '2020-01-02' },
+    { id: '4', title: '下午有空开个会', date: '2020-01-02' }
   ]
 
   private get buttonText() {
@@ -50,10 +50,18 @@ export default class Calender extends Vue {
     }
   }
 
-  private onClick() {
-    this.$modal.open({
-      component: CalenderDetail
-    })
+  private onEventClick({ event }) {
+    this.$modal.open(
+      CalenderDetail,
+      {
+        id: event.id,
+        title: event.title
+      },
+      {
+        title: '#' + event.id,
+        width: 980
+      }
+    )
   }
 }
 </script>