+++ src/shared/filters/index.ts
@@ -1,0 +1,5 @@
+import date from './date.filter'
+
+export default {
+  date
+}